requirements.txt
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
pyarrow>=14.0
